pyproject.toml
[build-system]
requires = ["poetry==0.12.15"]
build-backend = "poetry.masonry.api"


[tool.poetry]
name = "witness-angel-cryptolib"
description = "Witness Angel Cryptolib"
version = "0.1.0"
readme = "README.md"
authors = []

[tool.poetry.dependencies]
python = "^3.7.2"

# Cryptography
pycryptodome = "^3.9"
schema = "^0.7.0"
click = "^7.0"


[tool.poetry.dev-dependencies]

# Test runners and plugins
pytest = "^4.5"
pytest-cov = "^2.7"
pytest-randomly = "^3.0"
pytest-deadfixtures = "^2.1"
pytest-testmon = "^0.9"
pytest-timeout = "^1.3"

# Linters and formatters
black = {version = "^18.3-alpha.0",allows-prereleases = true}
pylint = "^2.3"
pylint-quotes = "^0.2.1"
safety = "^1.8"
yamllint = "^1.15"
xenon = "^0.5"
mypy = "^0.701"
doc8 = "^0.8"

# Documentation
sphinx = "^2.0"
sphinx-autodoc-typehints = "^1.6"
